+++ NEWS.rst
@@ -30,6 +30,8 @@ Changes in 3.18 (released ??/??/2011)
 *	The script ``doc2txt`` now reads from ``stdin`` and writes to ``stdout``
 	instead of requiring file names on the command line.
 
+*	If the script ``xml2xsc`` is called without arguments ``stdin`` is read.
+
 *	:mod:`ll.xist.ns.rest` now handles option lists.
 
 *	The Oracle URLs provided by :mod:`ll.orasql` now have a ``.sql`` extension